+++ source/Public/Get-ObjectNotFoundRecord.ps1
@@ -1,6 +1,6 @@
 <#
     .SYNOPSIS
-        Returns an invalid result exception object.
+        Returns an object not found exception object.
 
     .PARAMETER Message
         The message explaining why this error is being thrown.
@@ -9,7 +9,7 @@
         The error record containing the exception that is causing this terminating
         error.
 #>
-function Get-InvalidResultRecord
+function Get-ObjectNotFoundRecord
 {
     [CmdletBinding()]
     param